test: fix a second -Wdouble-promotion spot, caught by clang not GCC
Same pattern as the previous fix, one line up: clang's -Wdouble-promotion
also flags a plain implicit float-to-double conversion at a declaration
(not just in arithmetic), which GCC's build didn't catch but clang's did.
in[base + n] is float; make the cast explicit.

+++ tests/quality/test_search.cpp
@@ -218,7 +218,7 @@ TEST_CASE("searching on distortion lowers the decoded error", "[quality][search]
                 const auto& in = material[ch];
                 const std::size_t base = (frame - 1) * ac3::kSamplesPerFrame;
                 for (std::size_t n = 0; n < out.size(); ++n) {
-                    const double reference = in[base + n];
+                    const double reference = static_cast<double>(in[base + n]);
                     const double error = static_cast<double>(out[n]) - reference;
                     signal += reference * reference;
                     noise += error * error;
